refactor: 10단계 구현을 위한 패키지 구조 리팩토링
