+++ src/main/java/roomescape/core/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.core.domain;
 
 public class Reservation {
     private Long id;
